+++ src/main/java/com/cleaningsystem/dao/UserAccountDAO.java
@@ -5,7 +5,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 import com.cleaningsystem.model.UserAccount;
-import static com.cleaningsystem.dao.SQL_query.*;
+import static com.cleaningsystem.dao.Queries.*;
 import java.sql.ResultSet;
 import java.util.List;
 import java.sql.Date;
@@ -17,10 +17,10 @@ public class UserAccountDAO {
 
     private final RowMapper<UserAccount> userRowMapper = (ResultSet rs, int rowNum) -> {
         UserAccount user = new UserAccount();
-        user.setUID(rs.getInt("UID"));
+        user.setUid(rs.getInt("UID"));
         user.setName(rs.getString("name"));
         user.setAge(rs.getInt("age"));
-        user.setDOB(rs.getDate("dob").toLocalDate().toString());
+        user.setDob(rs.getDate("dob").toLocalDate().toString());
         user.setGender(rs.getString("gender"));
         user.setAddress(rs.getString("address"));
         user.setEmail(rs.getString("email"));
@@ -35,7 +35,7 @@ public class UserAccountDAO {
     }
 
     public int insertUserAccount(UserAccount user) {
-        java.sql.Date sqlDob = java.sql.Date.valueOf(user.getDOB());
+        java.sql.Date sqlDob = java.sql.Date.valueOf(user.getDob());
         return jdbcTemplate.update(CREATE_USER_ACCOUNT, 
             user.getName(), user.getAge(), sqlDob, user.getGender(), 
             user.getAddress(), user.getEmail(), user.getUsername(), user.getPassword(), 
@@ -55,9 +55,9 @@ public class UserAccountDAO {
 
     public boolean updateUserAccount(UserAccount user) {
         return jdbcTemplate.update(UPDATE_USER_ACCOUNT,
-            user.getName(), user.getAge(), user.getDOB(), user.getGender(),
+            user.getName(), user.getAge(), user.getDob(), user.getGender(),
             user.getAddress(), user.getEmail(), user.getUsername(), user.getProfileID(),
-            user.getUID()) > 0;
+            user.getUid()) > 0;
     }
 
     public boolean deleteUserAccount(int uid) {
