+++ src/main/java/com/cleaningsystem/dao/UserProfileDAO.java
@@ -5,7 +5,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 import com.cleaningsystem.model.UserProfile;
-import static com.cleaningsystem.dao.SQL_query.*;
+import static com.cleaningsystem.dao.Queries.*;
 import java.sql.ResultSet;
 import java.util.List;
