+++ src/main/resources/templates/dashboard.html
@@ -7,7 +7,7 @@
     <h1>Welcome to the Dashboard!</h1>
     <p>Name: <span th:text="*{name}"></span> </p>
     <p>Age: <span th:text="*{age}"></span> </p>
-    <!-- <p>DOB: <span th:text="*{dob}"></span> </p> -->
+    <p>DOB: <span th:text="*{dob}"></span> </p>
     <p>Gender: <span th:text="*{gender}"></span> </p>
     <p>Address: <span th:text="*{address}"></span> </p>
     <p>Email: <span th:text="*{email}"></span> </p>
