+++ src/main/resources/templates/signup.html
@@ -188,40 +188,46 @@
 
     <a href="index.html" class="back-btn">← Back</a>
 
-    <form>
+    <form th:action="@{/details}" method="post" th:object="${signUpForm}">
       <div class="form-row">
         <div class="form-group">
           <label for="fname">First name*</label>
-          <input type="text" id="fname" required>
+          <input type="text" id="fname" th:field="*{name}" required>
         </div>
-        <div class="form-group">
+        <!-- <div class="form-group">
           <label for="lname">Last name*</label>
           <input type="text" id="lname" required>
-        </div>
+        </div> -->
+        <select th:field="*{profileID}">
+          <option value="1">Platform Manager</option>
+          <option value="2">User Admin</option>
+          <option value="3">Home Owner</option>
+          <option value="4">Cleaner</option>
+      </select>
       </div>
 
       <div class="form-row">
         <div class="form-group">
           <label for="age">Age*</label>
-          <input type="number" id="age" min="18" required>
+          <input type="number" id="age" min="18" th:field="*{age}" required>
         </div>
         <div class="form-group">
           <label for="dob">Date of Birth*</label>
-          <input type="date" id="dob" required>
+          <input type="date" id="dob" th:field="*{dob}" required>
         </div>
       </div>
 
       <div class="form-row">
         <div class="form-group">
           <label for="address">Address*</label>
-          <input type="text" id="address" required>
+          <input type="text" id="address" th:field="*{address}" required>
         </div>
         <div class="form-group">
           <label>Gender*</label>
           <div class="gender-options">
-            <input type="radio" name="gender" value="Male" id="male" required><label for="male">Male</label>
-            <input type="radio" name="gender" value="Female" id="female"><label for="female">Female</label>
-            <input type="radio" name="gender" value="Other" id="other"><label for="other">Other</label>
+            <input type="radio" name="gender" value="Male" id="male" th:field="*{gender}"><label for="male">Male</label>
+            <input type="radio" name="gender" value="Female" id="female"><label for="female" th:field="*{gender}">Female</label>
+            <input type="radio" name="gender" value="Other" id="other"><label for="other" th:field="*{gender}">Other</label>
           </div>
         </div>
       </div>
@@ -229,21 +235,21 @@
       <div class="form-row">
         <div class="form-group full-width">
           <label for="email">Email address*</label>
-          <input type="email" id="email" required>
+          <input type="email" id="email" th:field="*{email}" required>
         </div>
       </div>
       
       <div class="form-row">
         <div class="form-group full-width">
           <label for="username">Username <small>(Only letter, numbers and underscores)</small></label>
-          <input type="text" id="username" pattern="^[a-zA-Z0-9_]+$" required>
+          <input type="text" id="username" pattern="^[a-zA-Z0-9_]+$" th:field="*{username}" required>
         </div>
       </div>
       
       <div class="form-row">
         <div class="form-group full-width">
           <label for="password">Password <small>(min. 8 char)</small></label>
-          <input type="password" id="password" minlength="8" required>
+          <input type="password" id="password" minlength="8" th:field="*{password}" required>
         </div>
       </div>
